Add files via upload
Movie Rating Prediction project involves building a model that predicts the rating of a movie based on features such as genre, director, and actors. By using regression techniques, the goal is to analyze historical movie data and develop a model that accurately estimates the ratings given to a movie by users or critics.

This project enables exploration of data analysis, preprocessing, feature engineering, and machine learning modeling techniques. It provides insights into the factors that influence movie ratings and helps in constructing a model that can estimate movie ratings with accuracy.